requirements.txt
voluptuous
pytest_voluptuous
curlify
allure-pytest
pytest~=7.1.2
selenium~=4.2.0
pytest~=7.1.3
selene>=2.0.b8
python-dotenv==0.20.0
Appium-Python-Client~=2.6.1
requests~=2.28.1
pydantic~=1.10.2